Add files via upload
In this project, I had built my first neural network and used it to predict daily bike rental ridership. The scope of the project is to save money of both customer and the dealer. This goal is achieved by creating a  simple neural network which used the dataset of previous years and predicted the future efficiently.